+++ lib/src/platform_handler/platform_handler_all.dart
@@ -20,11 +20,14 @@ class PlatformHandlerAll extends PlatformHandler {
       'Something went wrong, please report the issue https://www.github.com/incrediblezayed/file_saver/issues';
   late String directory = _somethingWentWrong;
 
-  final String _issueLink = 'https://www.github.com/incrediblezayed/file_saver/issues';
+  final String _issueLink =
+      'https://www.github.com/incrediblezayed/file_saver/issues';
 
   Future<String> saveFileForAndroid(FileModel fileModel) async {
     try {
-      directory = await _channel.invokeMethod<String>(_saveFile, fileModel.toMap()) ?? '';
+      directory =
+          await _channel.invokeMethod<String>(_saveFile, fileModel.toMap()) ??
+              '';
       return directory;
     } catch (e) {
       rethrow;
@@ -39,7 +42,8 @@ class PlatformHandlerAll extends PlatformHandler {
       throw Exception('The path was found null or empty');
     } else {
       final slash = Helpers.getFilePathSlash();
-      String filePath = '$path$slash${fileModel.name}${fileModel.ext}';
+      String filePath =
+          '$path$slash${fileModel.name}${fileModel.fileExtension}';
       final File file = File(filePath);
       await file.writeAsBytes(fileModel.bytes);
       bool exist = await file.exists();
@@ -68,7 +72,8 @@ class PlatformHandlerAll extends PlatformHandler {
     if (Platform.isAndroid || Platform.isIOS || Platform.isMacOS) {
       path = await _channel.invokeMethod<String>(_saveAs, fileModel.toMap());
     } else if (Platform.isWindows) {
-      final Int64List? bytes = await _channel.invokeMethod<Int64List?>('saveAs', fileModel.toMap());
+      final Int64List? bytes =
+          await _channel.invokeMethod<Int64List?>('saveAs', fileModel.toMap());
       path = bytes == null ? null : String.fromCharCodes(bytes);
     } else {
       throw UnimplementedError('Unimplemented Error');