+++ lib/src/utils/mime_types.dart
@@ -2,9 +2,6 @@ import 'package:collection/collection.dart';
 
 ///[MimeType] is an enum for adding filetype for HTML Blob
 enum MimeType {
-  ///[avi] for .avi extension
-  avi(name: 'AVI', type: 'video/x-msvideo'),
-
   ///[aac] for .aac extension
   aac(
     name: 'AAC',
@@ -22,6 +19,12 @@ enum MimeType {
     name: 'ASICS',
     type: 'application/vnd.etsi.asic-s+zip',
   ),
+
+  ///[avi] for .avi extension
+  avi(name: 'AVI', type: 'video/x-msvideo'),
+
+  ///[avif] for .avif extension
+  avif(name: 'AVIF', type: 'image/avif'),
 
   ///[bDoc] for .asice
   bDoc(
@@ -44,11 +47,20 @@ enum MimeType {
   ///[gif] for .gif extension
   gif(name: 'GIF', type: 'image/gif'),
 
-  ///[json] for .json extension
-  json(name: 'JSON', type: 'application/json'),
+  ///[heic] for .heic extension
+  heic(name: 'HEIC', type: 'image/heic'),
+
+  ///[heif] for .heif extension
+  heif(name: 'HEIF', type: 'image/heif'),
 
   ///[jpeg] for .jpeg extension
   jpeg(name: 'JPEG', type: 'image/jpeg'),
+
+  ///[json] for .json extension
+  json(name: 'JSON', type: 'application/json'),
+
+  ///[markdown] for .md extension
+  markdown(name: 'Markdown', type: 'text/markdown'),
 
   ///[microsoftExcel] for .xlsx extension
   microsoftExcel(
@@ -68,32 +80,20 @@ enum MimeType {
       type:
           'application/vnd.openxmlformats-officedocument.wordprocessingml.document'),
 
-  ///[mpeg] for .mpeg extension
-  mpeg(name: 'MPEG', type: 'video/mpeg'),
-
   ///[mp3] for .mp3 extension
   mp3(name: 'MP3', type: 'audio/mpeg'),
+
+  /// [mp4Audio] for .mp4 extension for audio files
+  mp4Audio(name: 'MP4 Audio', type: 'audio/mp4'),
+
+  /// [mp4Object] for .mp4 extension for other media files with mp4 extension
+  mp4Object(name: 'MP4 Object', type: 'application/mp4'),
 
   /// [mp4Video] for .mp4 extension for video files
   mp4Video(name: 'MP4 Video', type: 'video/mp4'),
 
-  /// [mp4Audio] for .mp4 extension for audio files
-  mp4Audio(name: 'MP4 Audio', type: 'audio/mp4'),
-  
-  /// [mp4Object] for .mp4 extension for other media files with mp4 extension
-  mp4Object(name: 'MP4 Object', type: 'application/mp4'),
-
-  ///[other] for other extension
-  other(name: 'Other', type: 'application/octet-stream'),
-
-  ///[otf] for .otf extension
-  otf(name: 'OTF', type: 'font/otf'),
-
-  ///[openDocSheets] for .ods extension
-  openDocSheets(
-    name: 'Open Document Sheets',
-    type: 'application/vnd.oasis.opendocument.spreadsheet',
-  ),
+  ///[mpeg] for .mpeg extension
+  mpeg(name: 'MPEG', type: 'video/mpeg'),
 
   ///[openDocPresentation] for .odp extension
   openDocPresentation(
@@ -101,10 +101,22 @@ enum MimeType {
     type: 'application/vnd.oasis.opendocument.presentation',
   ),
 
+  ///[openDocSheets] for .ods extension
+  openDocSheets(
+    name: 'Open Document Sheets',
+    type: 'application/vnd.oasis.opendocument.spreadsheet',
+  ),
+
   ///[openDocText] for .odt extension
   openDocText(
       name: 'Open Document Text',
       type: 'application/vnd.oasis.opendocument.text'),
+
+  ///[otf] for .otf extension
+  otf(name: 'OTF', type: 'font/otf'),
+
+  ///[other] for other extension
+  other(name: 'Other', type: 'application/octet-stream'),
 
   ///[pdf] for .pdf extension
   pdf(
@@ -121,11 +133,29 @@ enum MimeType {
     type: 'application/x-rar-compressed',
   ),
 
+  ///[sql] for .sql extension
+  sql(name: 'SQL', type: 'application/sql'),
+
+  ///[svg] for .svg extension
+  svg(name: 'SVG', type: 'image/svg+xml'),
+
   ///[text] for .txt extension
   text(name: 'Text', type: 'text/plain'),
 
   ///[ttf] for .ttf extension
   ttf(name: 'TTF', type: 'font/ttf'),
+
+  ///[webm] for .webm extension
+  webm(name: 'WebM', type: 'video/webm'),
+
+  ///[webp] for .webp extension
+  webp(name: 'WebP', type: 'image/webp'),
+
+  ///[xml] for .xml extension
+  xml(name: 'XML', type: 'application/xml'),
+
+  ///[yaml] for .yaml extension
+  yaml(name: 'YAML', type: 'application/x-yaml'),
 
   ///[zip] for .zip extension
   zip(